+++ CHANGELOG.md
@@ -6,6 +6,13 @@ stop printing references once they either have no reads mapped to them or
 the number of new reads mapped to them ("new" as in not already mapping to
 an earlier reference) falls to zero.
 
+## 4.0.4 July 7, 2021
+
+Added `fasta-split.py` command to split a FASTA/Q file into multiple files,
+each containing a given number of sequences. Added `dark/utils.py` function
+`take` to repeatedly yield lists of a given number of things from an
+iterable.
+
 ## 4.0.3 June 15, 2021
 
 Added `--topReferenceIdsFile` option to `sam-reference-read-counts.py` to